+++ core/inc/kversion.h
@@ -4,7 +4,7 @@
 /**                     RK0 — Real-Time Kernel '0'                            */
 /** Copyright (C) 2026 Antonio Giacomelli <dev@kernel0.org>                   */
 /**                                                                           */
-/** VERSION          :   V0.9.15                                              */
+/** VERSION          :   V0.9.16                                              */
 /** ARCHITECTURE     :   ARMv6/7M                                             */
 /**                                                                           */
 /** You may obtain a copy of the License at :                                 */
@@ -29,7 +29,7 @@ struct RK_gKversion
 
 #define RK_VERSION_MAJOR 0
 #define RK_VERSION_MINOR 9
-#define RK_VERSION_PATCH 15
+#define RK_VERSION_PATCH 16
 
 
 extern struct RK_gKversion const RK_gKversion;